erro ainda

+++ app/Http/Controllers/QuestionsController.php
@@ -40,8 +40,9 @@ class QuestionsController extends Controller
      */
     public function store(AskQuestionRequest $request)
     {
-       $request->user()->questions() ->create($request->only('title', 'body'));
-       return redirect()->route('questions.index')->with('success','Sua pergunta foi enviada');
+        $request->user()->questions()->create($request->only('title', 'body'));
+
+        return redirect()->route('questions.index')->with('success', "Your question has been submitted");
     }
 
     /**
@@ -52,9 +53,8 @@ class QuestionsController extends Controller
      */
     public function show(Question $question)
     {
-        $question->increment('views');//esta linha subst as duas abaixo
-        //$question->views = $question->views + 1;
-        //$question->save();
+        $question->increment('views');
+
         return view('questions.show', compact('question'));
     }
 
@@ -66,7 +66,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
+        return view("questions.edit", compact('question'));
     }
 
     /**
@@ -78,8 +81,13 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
+
         $question->update($request->only('title', 'body'));
-        return redirect('/questions')->with('success', 'Sua pergunta foi atualizada.');
+
+        return redirect('/questions')->with('success', "Your question has been updated.");
     }
 
     /**
@@ -90,7 +98,12 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, "Access denied");
+        }
+
         $question->delete();
-        return redirect('/questions')->with('success', 'Sua pergunta foi removida.');
+
+        return redirect('/questions')->with('success', "Your question has been deleted.");
     }
 }
